Fix empty relay state
Okta sends an empty relay state, if none is provided. This generates an
error, as the key is present in the post data, and by design different
from the current URL. So we try to redirect to the empty URL.

Also tuck the relaystate into a variable for readability and provide a
test for this case.